Change jerk recipe description and appearance
The section that indicated that pecan wood is hard to find in the US
has been removed. The site will be seen by people from other place
hence a more general description is better.

The paragraph is no longer a single line in the source code for ease of
reading.